[IMP] web: move and convert profiling_sweb_view
closes odoo/odoo#76871

X-original-commit: bee8953c3d8b7179bf1f526baa593825f09bfd97
