+++ addons/web/static/src/legacy/js/core/debug/profiling/profiling_qweb_view.scss
